refactor: streamline datepicker components and remove unused files
