--- react/datepicker/DatePickerNavigationProps.ts
@@ -1,7 +1,0 @@
-import type { Phrases } from './utils/defaultPhrases.js';
-
-export interface DatePickerNavigationProps {
-  phrases: Phrases;
-  nextMonth: () => void;
-  prevMonth: () => void;
-}
